Add some tests that `Pipeline.Start()` picks the right stdin/stdout
The most complicated code dealing with `Stage2` is the selection of
which types of stdin/stderr to pass to stages, and that's also the
main advantage of the `Stage2` interface. So add a bunch of tests that
the correct types (especially, `io.Pipe()` vs. `os.Pipe()`) are indeed
being selected.