Remove use of GOPROXY and its token
This is a public repository and therefore all of its dependencies should
be public.  It should not be necessary to use any internal proxy for
running tests or accessing Go binary because the default options should
be fine.

Remove the code to access private Go modules since it is unneeded.